Generalize Non_Band state to Error state to show various errors
